Level begin&end logic, signals and pause screen: An epic saga
Moved the code to show and hide menus and all relevant logic to the menu scripts instead of the game script

Made level_starts and level_ends scripts in banana_mouse, bubbles and pause_screen, so they can handle the necessary logic when a level begins and ends. For example process_mode and visibility

Added level_starts and level_ends signals in the game script. It connects the relevant children to them and emits them at the relevant times. Then the kids will do whatever is necessary to begin and end the game. Even if it means risking their lives and the lives of millions...