Goal fixed
Fixed the goal, which wasn't working properly before. Now the goal emits level_defeated, which causes the level to emit level_defeated. When starting a level, its level_defeated signal is connected to the _on_no_longer_in_level in the game script.

Fixed mouse visibility after defeating a level

Changed the names of the levels in the level_select